+++ src/api/AuthAPI.js
@@ -3,7 +3,14 @@ import axios from "axios";
 
 const AuthAPI = {
   async getSignIn(email, password) {
-    const data = await axios.post(`${BASE_URL}auth/signin`, {
+    const Data = await axios.post(`${BASE_URL}auth/signin`, {
+      email,
+      password,
+    });
+    return Data;
+  },
+  async getSignUp(email, password) {
+    const data = await axios.post(`${BASE_URL}auth/signup`, {
       email,
       password,
     });
