+++ src/components/AuthForm/AuthForm.style.jsx
@@ -1,9 +1,9 @@
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-import Button from "../../components/Button/Button";
+import Button from "../Button/Button";
 import { AllWrappCss, MainWrappCss } from "../../styles/Globalstyled";
 
-export const LoginWrapper = styled.section`
+export const InputFormWrapper = styled.section`
   ${AllWrappCss};
 `;
 export const MainWrapper = styled.section`
@@ -13,18 +13,18 @@ export const MainWrapper = styled.section`
   align-items: center;
 `;
 
-export const LoginTitle = styled.h1`
+export const MainTitle = styled.h1`
   margin: 30px 0 40px;
   font-weight: 500;
   font-size: 2.4rem;
   line-height: 3rem;
 `;
 
-export const LoginBtn = styled(Button)`
+export const SubmitBtn = styled(Button)`
   margin: 30px 0 20px;
 `;
 
-export const LoginLink = styled(Link)`
+export const JoinEmailLink = styled(Link)`
   font-size: 1.2rem;
   text-decoration: none;
   line-height: 1.5rem;
